# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Connector service, which handles any kind of external integrations for Dengankarya.",
        "title": "Connector API Docs",
        "contact": {
            "name": "Dengan Karya Support",
            "url": "https://dengankarya.com/contact",
            "email": "dukungan@dengankarya.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/shipments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a cursor-paginated list of shipments for the tenant. Supports filtering by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipping"
                ],
                "summary": "List shipments",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tenant ID",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor returned by previous response",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated statuses to filter (e.g. in_transit,delivered)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated shipment list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_dengankarya_connector_common.PaginationResponse-github_com_dengankarya_connector_internal_shipping_domain_Shipment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or cursor",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new draft shipment order via Biteship.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipping"
                ],
                "summary": "Create shipment",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tenant ID",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shipment creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.CreateShipmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Shipment created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_domain.Shipment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/shipments/couriers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of available shipping couriers from Biteship. Filter by comma-separated courier codes using the `only` query parameter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipping"
                ],
                "summary": "List couriers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated courier codes to filter (e.g. jne,sicepat)",
                        "name": "only",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Courier list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_dengankarya_connector_pkg_biteship.Courier"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/shipments/rates": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns available courier rates for a shipment before creating a draft order. Supply at least one location pair: area IDs (high accuracy), coordinates (required for instant couriers), or postal codes (easiest). Mix-and-match is supported (e.g. postal code origin + coordinate destination).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipping"
                ],
                "summary": "Get courier rates",
                "parameters": [
                    {
                        "description": "Rate query request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.GetRatesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available courier rates",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.GetRatesResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/shipments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a single shipment by its UUID for the authenticated tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipping"
                ],
                "summary": "Get shipment",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tenant ID",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Shipment UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shipment details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_domain.Shipment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid shipment ID",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "Shipment not found",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dengankarya_connector_common.PaginationResponse-github_com_dengankarya_connector_internal_shipping_domain_Shipment": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_domain.Shipment"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dengankarya_connector_internal_shipping_domain.Shipment": {
            "type": "object",
            "properties": {
                "confirmed_at": {
                    "type": "string"
                },
                "courier_code": {
                    "type": "string"
                },
                "courier_service_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_number": {
                    "description": "OrderNumber is the tokokarya's order number.",
                    "type": "string"
                },
                "picked_up_at": {
                    "type": "string"
                },
                "provider": {
                    "description": "Current default is `biteship`.",
                    "type": "string"
                },
                "provider_draft_order_id": {
                    "type": "string"
                },
                "provider_order_id": {
                    "type": "string"
                },
                "shipping_cost": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_domain.ShipmentStatus"
                },
                "tenant_id": {
                    "description": "TenantID is the merchant/tenant that owns this shipment.",
                    "type": "integer"
                },
                "tracking_number": {
                    "type": "string"
                },
                "tracking_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dengankarya_connector_internal_shipping_provider.CourierSelection": {
            "type": "object",
            "properties": {
                "company": {
                    "description": "Company is the courier provider code.\n\nExample values:\n  - jne\n  - sicepat\n  - anteraja",
                    "type": "string"
                },
                "type": {
                    "description": "Type is the courier service type.\n\nExample values:\n  - reg\n  - yes\n  - instant",
                    "type": "string"
                }
            }
        },
        "github_com_dengankarya_connector_internal_shipping_provider.CreateShipmentRequest": {
            "type": "object",
            "properties": {
                "courier": {
                    "description": "Courier contains the selected courier service.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.CourierSelection"
                        }
                    ]
                },
                "delivery": {
                    "description": "Delivery contains shipment delivery configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.DeliveryOptions"
                        }
                    ]
                },
                "destination": {
                    "description": "Destination contains receiver or delivery address information.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.Address"
                        }
                    ]
                },
                "items": {
                    "description": "Items contains the list of shipment items.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.ShipmentItem"
                    }
                },
                "metadata": {
                    "description": "Metadata contains arbitrary internal metadata.\n\nThis field may be forwarded to providers that support metadata storage.",
                    "type": "object",
                    "additionalProperties": {}
                },
                "note": {
                    "description": "Note contains additional shipment notes.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin contains pickup or sender address information.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.Address"
                        }
                    ]
                },
                "reference_id": {
                    "description": "ReferenceID is the internal order or shipment identifier.\n\nProviders may use this value as an idempotency key or external reference.",
                    "type": "string"
                },
                "shipper": {
                    "description": "Shipper contains optional shipper or merchant information.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.Shipper"
                        }
                    ]
                },
                "tags": {
                    "description": "Tags contains optional provider-level tagging information.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_dengankarya_connector_internal_shipping_provider.GetRatesRequest": {
            "type": "object",
            "properties": {
                "courier_insurance": {
                    "description": "CourierInsurance is the declared item value for insurance (optional).",
                    "type": "integer"
                },
                "couriers": {
                    "description": "Couriers is a comma-separated list of courier codes (required).",
                    "type": "string"
                },
                "destination_area_id": {
                    "type": "string"
                },
                "destination_cash_on_delivery": {
                    "description": "DestinationCashOnDelivery activates COD; cannot exceed IDR 15,000,000 (optional).",
                    "type": "integer"
                },
                "destination_cash_on_delivery_type": {
                    "description": "DestinationCODType is the COD disbursement window: \"3_days\", \"5_days\", or \"7_days\".",
                    "type": "string"
                },
                "destination_latitude": {
                    "type": "number"
                },
                "destination_longitude": {
                    "type": "number"
                },
                "destination_postal_code": {
                    "type": "integer"
                },
                "items": {
                    "description": "Items is the list of items to ship (required).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.ShipmentItem"
                    }
                },
                "origin_area_id": {
                    "description": "Area ID — highest accuracy; cannot be used with instant couriers.",
                    "type": "string"
                },
                "origin_latitude": {
                    "description": "Coordinates — required for instant couriers (Gojek, Grab, etc.).",
                    "type": "number"
                },
                "origin_longitude": {
                    "type": "number"
                },
                "origin_postal_code": {
                    "description": "Postal codes — medium accuracy, easiest to implement.",
                    "type": "integer"
                },
                "type": {
                    "description": "Type enables special rate modes (e.g. \"origin_suggestion_to_closest_destination\").",
                    "type": "string"
                }
            }
        },
        "github_com_dengankarya_connector_internal_shipping_provider.GetRatesResult": {
            "type": "object",
            "properties": {
                "destination": {
                    "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.RatesLocation"
                },
                "origin": {
                    "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.RatesLocation"
                },
                "pricing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.CourierRate"
                    }
                }
            }
        },
...
        "github_com_dengankarya_connector_pkg_biteship.Courier": {
            "type": "object",
            "properties": {
                "available_collection_method": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "available_for_cash_on_delivery": {
                    "type": "boolean"
                },
                "available_for_instant_waybill_id": {
                    "type": "boolean"
                },
                "available_for_proof_of_delivery": {
                    "type": "boolean"
                },
                "courier_code": {
                    "type": "string"
                },
                "courier_name": {
                    "type": "string"
                },
                "courier_service_code": {
                    "type": "string"
                },
                "courier_service_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "service_type": {
                    "type": "string"
                },
                "shipment_duration_range": {
                    "type": "string"
                },
                "shipment_duration_unit": {
                    "type": "string"
                },
                "shipping_type": {
                    "type": "string"
                },
                "tier": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    },
...
}